@@ -53,7 +53,7 @@
     "eslint-plugin-react-hooks": "^7.0.1",
     "eslint-plugin-simple-import-sort": "^12.1.1",
     "prettier": "^3.7.4",
-    "typescript": "~5.6.2",
+    "typescript": "~5.9.3",
     "typescript-eslint": "^8.49.0"
   },
   "engines": {
